Move stub eventer outside of tests and into dedicated package
